fix: revert rename of storage-web
