--- server/storage/storage-server/storage-entrance/src/main/resources/META-INF/spring.factories
@@ -1,1 +1,0 @@
-org.springframework.boot.autoconfigure.EnableAutoConfiguration=io.holoinsight.server.storage.entrance.HoloinsightStorageConfiguration
