+++ server/storage/storage-server/storage-web/src/main/java/io/holoinsight/server/storage/web/SlowSqlApi.java
@@ -1,0 +1,42 @@
+/*
+ * Copyright 2022 Holoinsight Project Authors. Licensed under Apache-2.0.
+ */
+package io.holoinsight.server.storage.web;
+
+import io.holoinsight.server.storage.common.model.query.QueryComponentRequest;
+import io.holoinsight.server.storage.common.model.query.SlowSql;
+import io.holoinsight.server.storage.web.model.FailResponse;
+import io.swagger.annotations.Api;
+import io.swagger.annotations.ApiOperation;
+import io.swagger.annotations.ApiParam;
+import io.swagger.annotations.ApiResponse;
+import io.swagger.annotations.ApiResponses;
+import io.swagger.annotations.Authorization;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RequestMethod;
+
+import javax.validation.Valid;
+import java.io.IOException;
+import java.util.List;
+
+@javax.annotation.Generated(value = "io.swagger.codegen.languages.SpringCodegen")
+@Api(value = "slowSql", description = "the slowSql API")
+@RequestMapping("/cluster/api/v1/slowSql")
+public interface SlowSqlApi {
+
+  @ApiOperation(value = "query slow sql list", nickname = "querySlowSqlList", notes = "查询慢sql列表",
+      response = SlowSql.class, authorizations = {@Authorization(value = "APIKeyHeader"),
+          @Authorization(value = "APIKeyQueryParam")},
+      tags = {"slowSqlList",})
+  @ApiResponses(value = {@ApiResponse(code = 200, message = "请求正常。", response = SlowSql.class),
+      @ApiResponse(code = 400, message = "请求失败。", response = FailResponse.class)})
+  @RequestMapping(value = "/query", produces = {"application/json"},
+      consumes = {"application/json"}, method = RequestMethod.POST)
+  ResponseEntity<List<SlowSql>> querySlowSqlList(
+      @ApiParam(value = "查询条件。", required = true) @Valid @RequestBody QueryComponentRequest request)
+      throws IOException;
+
+
+}
