+++ server/storage/storage-server/storage-web/src/main/resources/META-INF/spring.factories
@@ -1,0 +1,1 @@
+org.springframework.boot.autoconfigure.EnableAutoConfiguration=io.holoinsight.server.storage.web.HoloinsightStorageConfiguration
